requirements.txt
boto3==1.22.4
botocore==1.25.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
docopt==0.6.2
Flask==2.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysql-connector-repackaged==0.3.1
pipreqs==0.4.11
protobuf==3.20.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
yarg==0.1.9
